embeddings/pyproject.toml
[tool.poetry]
name = "nail-art-embeddings"
version = "0.1.0"
description = "CLIP embeddings and FAISS vector search for nail art images"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.3.0"
faiss-cpu = "^1.7.4"
numpy = "^1.24.3"
pillow = "^10.1.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
line-length = 88
target-version = "py39" 